a fixed length error rejecting comms protocol
The length is set by Arduino code and read into python on start up. Bad
data will be ignored but (and this needs to be fixed) it is currently
difficult for the Arduino to recover from comms error. it is probably
more difficult for an error to occur in normal operation though…